Restructure 0.1
