+++ src/App.jsx
@@ -1,13 +1,13 @@
 import './App.css'
-import Home from './Components/Home/Home';
+import Home from './Pages/Home/Home';
 import {Routes, Route} from 'react-router-dom'
-import SignIn from './Components/SignIn/SignIn';
-import SignUp from './Components/SignUp/SignUp';
-import Dasboard from './Components/Dasboard/Dasboard';
-import Profile from './Components/Profile/Profile';
-import UserProfile from './Components/UserProfile/UserProfile';
-import TournamentDescription from './Components/TournamentDescription/TournamentDescription';
-import Result from './Components/Result/Result';
+import SignIn from './Pages/SignIn/SignIn';
+import SignUp from './Pages/SignUp/SignUp';
+import Dashboard from './Pages/Dashboard/Dashboard';
+import Profile from './Pages/Profile/Profile';
+import UserProfile from './Pages/UserProfile/UserProfile';
+import TournamentDescription from './Pages/TournamentDescription/TournamentDescription';
+import Result from './Pages/Result/Result';
 function App() {
 
   return (
@@ -18,7 +18,7 @@ function App() {
         <Route exact path="/" element={<Home/>}></Route>
         <Route exact path="/signin" element={<SignIn/>}></Route>
         <Route exact path="/signup" element={<SignUp/>}></Route>
-        <Route exact path="/dashboard" element={<Dasboard/>}></Route>
+        <Route exact path="/dashboard" element={<Dashboard/>}></Route>
         <Route exact path="/profile" element={<Profile/>}></Route>
         <Route exact path="/profile/edit" element={<UserProfile/>}></Route>   
         <Route exact path="/tournaments/details" element={<TournamentDescription/>}></Route>
