+++ src/Pages/Dashboard/Dashboard.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Navbar from '../Navbar/Navbar'
+import Navbar from '../../Components/Navbar/Navbar'
 
 export default function Dasboard() {
   return (
